+++ src/routes/computePlanDetails/components/Actions.tsx
@@ -1,8 +1,17 @@
-import { Box, HStack, IconButton, Tooltip } from '@chakra-ui/react';
-import { RiStarLine, RiStarFill } from 'react-icons/ri';
+import {
+    Box,
+    HStack,
+    IconButton,
+    Menu,
+    MenuButton,
+    MenuList,
+    Tooltip,
+} from '@chakra-ui/react';
+import { RiStarLine, RiStarFill, RiMoreLine } from 'react-icons/ri';
 
 import useFavoriteComputePlans from '@/hooks/useFavoriteComputePlans';
 import { ComputePlanT } from '@/modules/computePlans/ComputePlansTypes';
+import useCancelComputePlan from '@/routes/computePlanDetails/hooks/useCancelComputePlan';
 
 import PerfDownloadButton from '@/components/PerfDownloadButton';
 
@@ -20,6 +29,9 @@ const Actions = ({
 
     const favorite = !loading && computePlan && isFavorite(computePlan.key);
     const ariaLabel = favorite ? 'Remove from favorites' : 'Add to favorites';
+
+    const { cancelComputePlanDialog, cancelComputePlanMenuItem } =
+        useCancelComputePlan(computePlan);
 
     return (
         <HStack paddingX="8">
@@ -47,6 +59,21 @@ const Actions = ({
             <Box>
                 <PerfDownloadButton />
             </Box>
+            {cancelComputePlanDialog && cancelComputePlanMenuItem && (
+                <Box>
+                    <Menu>
+                        <MenuButton
+                            as={IconButton}
+                            aria-label="Options"
+                            icon={<RiMoreLine />}
+                            variant="outline"
+                            size="xs"
+                        />
+                        <MenuList>{cancelComputePlanMenuItem}</MenuList>
+                    </Menu>
+                    {cancelComputePlanDialog}
+                </Box>
+            )}
         </HStack>
     );
 };